backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
httpx==0.26.0
openai==1.10.0
tenacity==8.2.3
orjson==3.9.12
slowapi==0.1.9
python-dotenv==1.0.0
pymongo==4.6.1
pdfplumber==0.11.0
pypdf==3.17.4

# 문서 업로드 및 RAG 기능 추가 패키지
PyPDF2>=3.0.1
python-docx>=1.1.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1
langchain-openai>=0.0.1
faiss-cpu>=1.7.4
